# backend/docs/swagger.yaml
basePath: /
definitions:
...
  consts.CopySetting:
    enum:
    - ""
    - append
    - disabled
    type: string
    x-enum-comments:
      CopySettingAppend: 增加内容尾巴
      CopySettingDisabled: 禁止复制内容
      CopySettingNone: 无限制
    x-enum-descriptions:
    - 无限制
    - 增加内容尾巴
    - 禁止复制内容
    x-enum-varnames:
    - CopySettingNone
    - CopySettingAppend
    - CopySettingDisabled
...
  consts.HomePageSetting:
    enum:
    - doc
    - custom
    type: string
    x-enum-comments:
      HomePageSettingCustom: 自定义首页
      HomePageSettingDoc: 文档页面
    x-enum-descriptions:
    - 文档页面
    - 自定义首页
    x-enum-varnames:
    - HomePageSettingDoc
    - HomePageSettingCustom
...
  consts.NodeAccessPerm:
    enum:
    - open
    - partial
    - closed
    type: string
    x-enum-comments:
      NodeAccessPermClosed: 完全禁止
      NodeAccessPermOpen: 完全开放
      NodeAccessPermPartial: 部分开放
    x-enum-descriptions:
    - 完全开放
    - 部分开放
    - 完全禁止
    x-enum-varnames:
    - NodeAccessPermOpen
    - NodeAccessPermPartial
    - NodeAccessPermClosed
...
  consts.WatermarkSetting:
    enum:
    - ""
    - hidden
    - visible
    type: string
    x-enum-comments:
      WatermarkDisabled: 未开启水印
      WatermarkHidden: 隐形水印
      WatermarkVisible: 显性水印
    x-enum-descriptions:
    - 未开启水印
    - 隐形水印
    - 显性水印
    x-enum-varnames:
    - WatermarkDisabled
    - WatermarkHidden
    - WatermarkVisible
  domain.AIFeedbackSettings:
    properties:
      ai_feedback_type:
        items:
          type: string
        type: array
      is_enabled:
        type: boolean
    type: object
...
  domain.AppDetailResp:
    properties:
      id:
        type: string
      kb_id:
        type: string
      name:
        type: string
      recommend_nodes:
        items:
          $ref: '#/definitions/domain.RecommendNodeListResp'
        type: array
      settings:
        $ref: '#/definitions/domain.AppSettingsResp'
      type:
        $ref: '#/definitions/domain.AppType'
    type: object
...
  domain.AppSettings:
    properties:
      ai_feedback_settings:
        allOf:
        - $ref: '#/definitions/domain.AIFeedbackSettings'
        description: AI feedback
      body_code:
        type: string
      btns:
        items: {}
        type: array
      catalog_settings:
        allOf:
        - $ref: '#/definitions/domain.CatalogSettings'
        description: catalog settings
      contribute_settings:
        $ref: '#/definitions/domain.ContributeSettings'
      conversation_setting:
        $ref: '#/definitions/domain.ConversationSetting'
      copy_setting:
        allOf:
        - $ref: '#/definitions/consts.CopySetting'
        enum:
        - ""
        - append
        - disabled
      desc:
        description: seo
        type: string
      dingtalk_bot_client_id:
        type: string
      dingtalk_bot_client_secret:
        type: string
      dingtalk_bot_is_enabled:
        description: DingTalkBot
        type: boolean
      dingtalk_bot_template_id:
        type: string
      disclaimer_settings:
        allOf:
        - $ref: '#/definitions/domain.DisclaimerSettings'
        description: Disclaimer Settings
      discord_bot_is_enabled:
        description: DisCordBot
        type: boolean
      discord_bot_token:
        type: string
      document_feedback_is_enabled:
        description: document feedback
        type: boolean
      feishu_bot_app_id:
        type: string
      feishu_bot_app_secret:
        type: string
      feishu_bot_is_enabled:
        description: FeishuBot
        type: boolean
      footer_settings:
        allOf:
        - $ref: '#/definitions/domain.FooterSettings'
        description: footer settings
      head_code:
        description: inject code
        type: string
      home_page_setting:
        $ref: '#/definitions/consts.HomePageSetting'
      icon:
        type: string
      keyword:
        type: string
      lark_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.LarkBotSettings'
        description: LarkBot
      mcp_server_settings:
        allOf:
        - $ref: '#/definitions/domain.MCPServerSettings'
        description: MCP Server Settings
      openai_api_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.OpenAIAPIBotSettings'
        description: OpenAI API Bot settings
      recommend_node_ids:
        items:
          type: string
        type: array
      recommend_questions:
        items:
          type: string
        type: array
      search_placeholder:
        type: string
      stats_setting:
        $ref: '#/definitions/domain.StatsSetting'
      theme_and_style:
        $ref: '#/definitions/domain.ThemeAndStyle'
      theme_mode:
        description: theme
        type: string
      title:
        description: nav
        type: string
      watermark_content:
        type: string
      watermark_setting:
        allOf:
        - $ref: '#/definitions/consts.WatermarkSetting'
        enum:
        - ""
        - hidden
        - visible
      web_app_comment_settings:
        allOf:
        - $ref: '#/definitions/domain.WebAppCommentSettings'
        description: webapp comment settings
      web_app_custom_style:
        allOf:
        - $ref: '#/definitions/domain.WebAppCustomSettings'
        description: WebAppCustomStyle
      web_app_landing_configs:
        description: WebAppLandingConfigs
        items:
          $ref: '#/definitions/domain.WebAppLandingConfig'
        type: array
      web_app_landing_theme:
        $ref: '#/definitions/domain.WebAppLandingTheme'
      wechat_app_advanced_setting:
        $ref: '#/definitions/domain.WeChatAppAdvancedSetting'
      wechat_app_agent_id:
        type: string
      wechat_app_corpid:
        type: string
      wechat_app_encodingaeskey:
        type: string
      wechat_app_is_enabled:
        description: WechatAppBot 企业微信机器人
        type: boolean
      wechat_app_secret:
        type: string
      wechat_app_token:
        type: string
      wechat_official_account_app_id:
        type: string
      wechat_official_account_app_secret:
        type: string
      wechat_official_account_encodingaeskey:
        type: string
      wechat_official_account_is_enabled:
        description: WechatOfficialAccount
        type: boolean
      wechat_official_account_token:
        type: string
      wechat_service_contain_keywords:
        items:
          type: string
        type: array
      wechat_service_corpid:
        type: string
      wechat_service_encodingaeskey:
        type: string
      wechat_service_equal_keywords:
        items:
          type: string
        type: array
      wechat_service_is_enabled:
        description: WechatServiceBot
        type: boolean
      wechat_service_secret:
        type: string
      wechat_service_token:
        type: string
      wecom_ai_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.WecomAIBotSettings'
        description: WecomAIBotSettings 企业微信智能机器人
      welcome_str:
        description: welcome
        type: string
      widget_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.WidgetBotSettings'
        description: Widget bot settings
    type: object
  domain.AppSettingsResp:
    properties:
      ai_feedback_settings:
        allOf:
        - $ref: '#/definitions/domain.AIFeedbackSettings'
        description: AI feedback
      body_code:
        type: string
      btns:
        items: {}
        type: array
      catalog_settings:
        allOf:
        - $ref: '#/definitions/domain.CatalogSettings'
        description: catalog settings
      contribute_settings:
        $ref: '#/definitions/domain.ContributeSettings'
      conversation_setting:
        $ref: '#/definitions/domain.ConversationSetting'
      copy_setting:
        $ref: '#/definitions/consts.CopySetting'
      desc:
        description: seo
        type: string
      dingtalk_bot_client_id:
        type: string
      dingtalk_bot_client_secret:
        type: string
      dingtalk_bot_is_enabled:
        description: DingTalkBot
        type: boolean
      dingtalk_bot_template_id:
        type: string
      disclaimer_settings:
        allOf:
        - $ref: '#/definitions/domain.DisclaimerSettings'
        description: Disclaimer Settings
      discord_bot_is_enabled:
        description: DisCordBot
        type: boolean
      discord_bot_token:
        type: string
      document_feedback_is_enabled:
        description: document feedback
        type: boolean
      feishu_bot_app_id:
        type: string
      feishu_bot_app_secret:
        type: string
      feishu_bot_is_enabled:
        description: FeishuBot
        type: boolean
      footer_settings:
        allOf:
        - $ref: '#/definitions/domain.FooterSettings'
        description: footer settings
      head_code:
        description: inject code
        type: string
      home_page_setting:
        $ref: '#/definitions/consts.HomePageSetting'
      icon:
        type: string
      keyword:
        type: string
      lark_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.LarkBotSettings'
        description: LarkBot
      mcp_server_settings:
        allOf:
        - $ref: '#/definitions/domain.MCPServerSettings'
        description: MCP Server Settings
      openai_api_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.OpenAIAPIBotSettings'
        description: OpenAI API settings
      recommend_node_ids:
        items:
          type: string
        type: array
      recommend_questions:
        items:
          type: string
        type: array
      search_placeholder:
        type: string
      stats_setting:
        $ref: '#/definitions/domain.StatsSetting'
      theme_and_style:
        $ref: '#/definitions/domain.ThemeAndStyle'
      theme_mode:
        description: theme
        type: string
      title:
        description: nav
        type: string
      watermark_content:
        type: string
      watermark_setting:
        $ref: '#/definitions/consts.WatermarkSetting'
      web_app_comment_settings:
        allOf:
        - $ref: '#/definitions/domain.WebAppCommentSettings'
        description: webapp comment settings
      web_app_custom_style:
        allOf:
        - $ref: '#/definitions/domain.WebAppCustomSettings'
        description: WebAppCustomStyle
      web_app_landing_configs:
        description: WebApp Landing Settings
        items:
          $ref: '#/definitions/domain.WebAppLandingConfigResp'
        type: array
      web_app_landing_theme:
        $ref: '#/definitions/domain.WebAppLandingTheme'
      wechat_app_advanced_setting:
        $ref: '#/definitions/domain.WeChatAppAdvancedSetting'
      wechat_app_agent_id:
        type: string
      wechat_app_corpid:
        type: string
      wechat_app_encodingaeskey:
        type: string
      wechat_app_is_enabled:
        description: WechatAppBot
        type: boolean
      wechat_app_secret:
        type: string
      wechat_app_token:
        type: string
      wechat_official_account_app_id:
        type: string
      wechat_official_account_app_secret:
        type: string
      wechat_official_account_encodingaeskey:
        type: string
      wechat_official_account_is_enabled:
        description: WechatOfficialAccount
        type: boolean
      wechat_official_account_token:
        type: string
      wechat_service_contain_keywords:
        items:
          type: string
        type: array
      wechat_service_corpid:
        type: string
      wechat_service_encodingaeskey:
        type: string
      wechat_service_equal_keywords:
        items:
          type: string
        type: array
      wechat_service_is_enabled:
        description: WechatServiceBot
        type: boolean
      wechat_service_secret:
        type: string
      wechat_service_token:
        type: string
      wecom_ai_bot_settings:
        $ref: '#/definitions/domain.WecomAIBotSettings'
      welcome_str:
        description: welcome
        type: string
      widget_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.WidgetBotSettings'
        description: WidgetBot
    type: object
  domain.AppType:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    format: int32
    type: integer
    x-enum-varnames:
    - AppTypeWeb
    - AppTypeWidget
    - AppTypeDingTalkBot
    - AppTypeFeishuBot
    - AppTypeWechatBot
    - AppTypeWechatServiceBot
    - AppTypeDisCordBot
    - AppTypeWechatOfficialAccount
    - AppTypeOpenAIAPI
    - AppTypeWecomAIBot
    - AppTypeLarkBot
    - AppTypeMcpServer
...
  domain.BannerConfig:
    properties:
      bg_url:
        type: string
      btns:
        items:
          properties:
            href:
              type: string
            id:
              type: string
            text:
              type: string
            type:
              type: string
          type: object
        type: array
      hot_search:
        items:
          type: string
        type: array
      placeholder:
        type: string
      subtitle:
        type: string
      subtitle_color:
        type: string
      subtitle_font_size:
        type: integer
      title:
        type: string
      title_color:
        type: string
      title_font_size:
        type: integer
    type: object
  domain.BasicDocConfig:
    properties:
      bg_color:
        type: string
      title:
        type: string
      title_color:
        type: string
    type: object
...
  domain.BlockGridConfig:
    properties:
      list:
        items:
          properties:
            id:
              type: string
            name:
              type: string
            url:
              type: string
          type: object
        type: array
      title:
        type: string
      type:
        type: string
    type: object
  domain.BrandGroup:
    properties:
      links:
        items:
          $ref: '#/definitions/domain.Link'
        type: array
      name:
        type: string
    type: object
...
  domain.CarouselConfig:
    properties:
      bg_color:
        type: string
      list:
        items:
          properties:
            desc:
              type: string
            id:
              type: string
            title:
              type: string
            url:
              type: string
          type: object
        type: array
      title:
        type: string
    type: object
  domain.CaseConfig:
    properties:
      list:
        items:
          properties:
            id:
              type: string
            link:
              type: string
            name:
              type: string
          type: object
        type: array
      title:
        type: string
      type:
        type: string
    type: object
  domain.CatalogSettings:
    properties:
      catalog_folder:
        description: '1: 展开, 2: 折叠, default: 1'
        type: integer
      catalog_visible:
        description: '1: 显示, 2: 隐藏, default: 1'
        type: integer
      catalog_width:
        description: '200 - 300, default: 260'
        type: integer
    type: object
...
  domain.CommentConfig:
    properties:
      list:
        items:
          properties:
            avatar:
              type: string
            comment:
              type: string
            id:
              type: string
            profession:
              type: string
            user_name:
              type: string
          type: object
        type: array
      title:
        type: string
      type:
        type: string
    type: object
...
  domain.ContributeSettings:
    properties:
      is_enable:
        type: boolean
    type: object
...
  domain.ConversationSetting:
    properties:
      copyright_hide_enabled:
        type: boolean
      copyright_info:
        type: string
    type: object
...
  domain.DirDocConfig:
    properties:
      bg_color:
        type: string
      title:
        type: string
      title_color:
        type: string
    type: object
  domain.DisclaimerSettings:
    properties:
      content:
        type: string
    type: object
...
  domain.FaqConfig:
    properties:
      bg_color:
        type: string
      list:
        items:
          properties:
            id:
              type: string
            link:
              type: string
            question:
              type: string
          type: object
        type: array
      title:
        type: string
      title_color:
        type: string
    type: object
  domain.FeatureConfig:
    properties:
      list:
        items:
          properties:
            desc:
              type: string
            id:
              type: string
            name:
              type: string
          type: object
        type: array
      title:
        type: string
      type:
        type: string
    type: object
...
  domain.FooterSettings:
    properties:
      brand_desc:
        type: string
      brand_groups:
        items:
          $ref: '#/definitions/domain.BrandGroup'
        type: array
      brand_logo:
        type: string
      brand_name:
        type: string
      corp_name:
        type: string
      footer_style:
        type: string
      icp:
        type: string
    type: object
...
  domain.ImgTextConfig:
    properties:
      item:
        properties:
          desc:
            type: string
          name:
            type: string
          url:
            type: string
        type: object
      title:
        type: string
      type:
        type: string
    type: object
...
  domain.LarkBotSettings:
    properties:
      app_id:
        type: string
      app_secret:
        type: string
      encrypt_key:
        type: string
      is_enabled:
        type: boolean
      verify_token:
        type: string
    type: object
  domain.Link:
    properties:
      name:
        type: string
      url:
        type: string
    type: object
  domain.MCPServerSettings:
    properties:
      docs_tool_settings:
        $ref: '#/definitions/domain.MCPToolSettings'
      is_enabled:
        type: boolean
      sample_auth:
        $ref: '#/definitions/domain.SimpleAuth'
    type: object
  domain.MCPToolSettings:
    properties:
      desc:
        type: string
      name:
        type: string
    type: object
...
  domain.MetricsConfig:
    properties:
      list:
        items:
          properties:
            id:
              type: string
            name:
              type: string
            number:
              type: string
          type: object
        type: array
      title:
        type: string
      type:
        type: string
    type: object
...
  domain.NodePermissions:
    properties:
      answerable:
        allOf:
        - $ref: '#/definitions/consts.NodeAccessPerm'
        description: 可被问答
      visible:
        allOf:
        - $ref: '#/definitions/consts.NodeAccessPerm'
        description: 导航内可见
      visitable:
        allOf:
        - $ref: '#/definitions/consts.NodeAccessPerm'
        description: 可被访问
    type: object
...
  domain.NodeType:
    enum:
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - NodeTypeFolder
    - NodeTypeDocument
...
  domain.OpenAIAPIBotSettings:
    properties:
      is_enabled:
        type: boolean
      secret_key:
        type: string
    type: object
...
  domain.PWResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.QuestionConfig:
    properties:
      list:
        items:
          properties:
            id:
              type: string
            question:
              type: string
          type: object
        type: array
      title:
        type: string
      type:
        type: string
    type: object
...
  domain.RecommendNodeListResp:
    properties:
      emoji:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      position:
        type: number
      recommend_nodes:
        items:
          $ref: '#/definitions/domain.RecommendNodeListResp'
        type: array
      summary:
        type: string
      type:
        $ref: '#/definitions/domain.NodeType'
    type: object
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.SimpleAuth:
    properties:
      enabled:
        type: boolean
      password:
        type: string
    type: object
  domain.SimpleDocConfig:
    properties:
      bg_color:
        type: string
      title:
        type: string
      title_color:
        type: string
    type: object
  domain.SocialMediaAccount:
    properties:
      channel:
        type: string
      icon:
        type: string
      link:
        type: string
      phone:
        type: string
      text:
        type: string
    type: object
...
  domain.StatsSetting:
    properties:
      pv_enable:
        type: boolean
    type: object
...
  domain.TextConfig:
    properties:
      title:
        type: string
      type:
        type: string
    type: object
  domain.TextImgConfig:
    properties:
      item:
        properties:
          desc:
            type: string
          name:
            type: string
          url:
            type: string
        type: object
      title:
        type: string
      type:
        type: string
    type: object
...
  domain.ThemeAndStyle:
    properties:
      bg_image:
        type: string
      doc_width:
        type: string
    type: object
  domain.UpdateAppReq:
    properties:
      kb_id:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/domain.AppSettings'
    type: object
...
  domain.WeChatAppAdvancedSetting:
    properties:
      disclaimer_content:
        type: string
      feedback_enable:
        type: boolean
      feedback_type:
        items:
          type: string
        type: array
      prompt:
        type: string
      text_response_enable:
        type: boolean
    type: object
  domain.WebAppCommentSettings:
    properties:
      is_enable:
        type: boolean
      moderation_enable:
        type: boolean
    type: object
  domain.WebAppCustomSettings:
    properties:
      allow_theme_switching:
        type: boolean
      footer_show_intro:
        type: boolean
      header_search_placeholder:
        type: string
      show_brand_info:
        type: boolean
      social_media_accounts:
        items:
          $ref: '#/definitions/domain.SocialMediaAccount'
        type: array
    type: object
  domain.WebAppLandingConfig:
    properties:
      banner_config:
        $ref: '#/definitions/domain.BannerConfig'
      basic_doc_config:
        $ref: '#/definitions/domain.BasicDocConfig'
      block_grid_config:
        $ref: '#/definitions/domain.BlockGridConfig'
      carousel_config:
        $ref: '#/definitions/domain.CarouselConfig'
      case_config:
        $ref: '#/definitions/domain.CaseConfig'
      com_config_order:
        items:
          type: string
        type: array
      comment_config:
        $ref: '#/definitions/domain.CommentConfig'
      dir_doc_config:
        $ref: '#/definitions/domain.DirDocConfig'
      faq_config:
        $ref: '#/definitions/domain.FaqConfig'
      feature_config:
        $ref: '#/definitions/domain.FeatureConfig'
      img_text_config:
        $ref: '#/definitions/domain.ImgTextConfig'
      metrics_config:
        $ref: '#/definitions/domain.MetricsConfig'
      node_ids:
        items:
          type: string
        type: array
      question_config:
        $ref: '#/definitions/domain.QuestionConfig'
      simple_doc_config:
        $ref: '#/definitions/domain.SimpleDocConfig'
      text_config:
        $ref: '#/definitions/domain.TextConfig'
      text_img_config:
        $ref: '#/definitions/domain.TextImgConfig'
      type:
        type: string
    type: object
  domain.WebAppLandingConfigResp:
    properties:
      banner_config:
        $ref: '#/definitions/domain.BannerConfig'
      basic_doc_config:
        $ref: '#/definitions/domain.BasicDocConfig'
      block_grid_config:
        $ref: '#/definitions/domain.BlockGridConfig'
      carousel_config:
        $ref: '#/definitions/domain.CarouselConfig'
      case_config:
        $ref: '#/definitions/domain.CaseConfig'
      com_config_order:
        items:
          type: string
        type: array
      comment_config:
        $ref: '#/definitions/domain.CommentConfig'
      dir_doc_config:
        $ref: '#/definitions/domain.DirDocConfig'
      faq_config:
        $ref: '#/definitions/domain.FaqConfig'
      feature_config:
        $ref: '#/definitions/domain.FeatureConfig'
      img_text_config:
        $ref: '#/definitions/domain.ImgTextConfig'
      metrics_config:
        $ref: '#/definitions/domain.MetricsConfig'
      node_ids:
        items:
          type: string
        type: array
      nodes:
        items:
          $ref: '#/definitions/domain.RecommendNodeListResp'
        type: array
      question_config:
        $ref: '#/definitions/domain.QuestionConfig'
      simple_doc_config:
        $ref: '#/definitions/domain.SimpleDocConfig'
      text_config:
        $ref: '#/definitions/domain.TextConfig'
      text_img_config:
        $ref: '#/definitions/domain.TextImgConfig'
      type:
        type: string
    type: object
  domain.WebAppLandingTheme:
    properties:
      name:
        type: string
    type: object
  domain.WecomAIBotSettings:
    properties:
      encodingaeskey:
        type: string
      is_enabled:
        type: boolean
      token:
        type: string
    type: object
  domain.WidgetBotSettings:
    properties:
      btn_id:
        type: string
      btn_logo:
        type: string
      btn_position:
        type: string
      btn_style:
        type: string
      btn_text:
        type: string
      copyright_hide_enabled:
        type: boolean
      copyright_info:
        type: string
      disclaimer:
        type: string
      is_open:
        type: boolean
      modal_position:
        type: string
      placeholder:
        type: string
      recommend_node_ids:
        items:
          type: string
        type: array
      recommend_questions:
        items:
          type: string
        type: array
      search_mode:
        type: string
      theme_mode:
        type: string
    type: object
...
info:
  contact: {}
  description: panda-wiki API documentation
  title: panda-wiki API
  version: "1.0"
paths:
  /api/v1/app:
    delete:
      consumes:
      - application/json
      description: Delete app
      parameters:
      - description: kb id
        in: query
        name: kb_id
        required: true
        type: string
      - description: app id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Delete app
      tags:
      - app
    put:
      consumes:
      - application/json
      description: Update app
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: app
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Update app
      tags:
      - app
  /api/v1/app/detail:
    get:
      consumes:
      - application/json
      description: Get app detail
      parameters:
      - description: kb id
        in: query
        name: kb_id
        required: true
        type: string
      - description: app type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.AppDetailResp'
              type: object
      security:
      - bearerAuth: []
      summary: Get app detail
      tags:
      - app
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"